requirements.txt
fastapi
uvicorn
python-multipart
python-jose
passlib[bcrypt]
pypdf
python-docx
openpyxl
